Return shapes to primitives_2d
