+++ pyprocessing/__init__.py
@@ -18,11 +18,10 @@ from pyprocessing.environment import (
     size, frame_rate, delay,
 )
 from pyprocessing.primitives_2d import (
-    ellipse, rect, triangle, quad, arc
-)
-from pyprocessing.shapes_2d import(
+    ellipse, rect, triangle, quad, arc,
     square, circle, line
 )
+
 from pyprocessing.color import (
     Color, alpha, red, green, blue,
     brightness, hue, saturation, lerp_color,
