+++ pyprocessing/primitives_2d.py
@@ -35,3 +35,18 @@ def quad(x1, y1, x2, y2, x3, y3, x4, y4):
 def arc(x, y, width, height, start, stop, mode='PIE'):
     global pp
     pp.windows.set_arc(x, y, width, height, start, stop, mode)
+
+
+def circle(x, y, size):
+    global pp
+    pp.windows.set_ellipse(x, y, size, size)
+
+
+def square(x, y, extent):
+    global pp
+    pp.windows.set_rectangle(x, y, extent, extent)
+
+
+def line(x1, y1, x2, y2):
+    global pp
+    pp.windows.set_line(x1, y1, x2, y2)
